Reverse merging 319797-319808
git-svn-id: https://source.sakaiproject.org/svn/profile2/branches/sakai-10.x@320035 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ tool/src/java/org/sakaiproject/profile2/tool/pages/ViewPictures.java
@@ -73,7 +73,7 @@ public class ViewPictures extends BasePage {
 
 		IDataProvider dataProvider = new GalleryImageDataProvider(userUuid);
 
-		long numImages = dataProvider.size();
+		int numImages = dataProvider.size();
 		
 		gridView = new GridView("rows", dataProvider) {
